requirements.txt
python-telegram-bot[webhooks]==20.7
asyncpg==0.29.0
anthropic>=0.40.0
python-dotenv==1.0.1
aiohttp>=3.9.0
matplotlib>=3.8.0
apscheduler>=3.10.0
